@@ -27,7 +27,7 @@ class Element:
         pass
 
     def potinf(self, x, y, aq=None):
-        """Returns array of size (nparam, naq)"""
+        """Returns array of size (nparam, naq)."""
         raise Exception("Must overload Element.potinf()")
 
     def potential(self, x, y, aq=None):
@@ -36,8 +36,10 @@ class Element:
         return np.sum(self.parameters * self.potinf(x, y, aq), 0)
 
     def potinflayers(self, x, y, layers, aq=None):
-        """Returns array of size (len(layers),nparam) only used in building
-        equations."""
+        """Returns array of size (len(layers),nparam).
+
+        Only used in building equations.
+        """
         if aq is None:
             aq = self.model.aq.find_aquifer_data(x, y)
         pot = self.potinf(x, y, aq)  # nparam rows, naq cols
@@ -54,18 +56,20 @@ class Element:
         return pot[layers]
 
     def disvecinf(self, x, y, aq=None):
-        """Returns array of size (2, nparam, naq)"""
+        """Returns array of size (2, nparam, naq)."""
         raise Exception("Must overload Element.disvecinf()")
 
     def disvec(self, x, y, aq=None):
-        """Returns array of size (2, nparam, naq)"""
+        """Returns array of size (2, nparam, naq)."""
         if aq is None:
             aq = self.model.aq.find_aquifer_data(x, y)
         return np.sum(self.parameters * self.disvecinf(x, y, aq), 1)
 
     def disvecinflayers(self, x, y, layers, aq=None):
-        """Returns two arrays of size (len(layers),nparam) only used in building
-        equations."""
+        """Returns two arrays of size (len(layers),nparam).
+
+        Only used in building equations.
+        """
         if aq is None:
             aq = self.model.aq.find_aquifer_data(x, y)
         qxqy = self.disvecinf(x, y, aq)  # nparam rows, naq cols
