@@ -47,7 +47,7 @@ class PolygonInhom(AquiferData):
         self.ymax = max(self.y)
 
     def __repr__(self):
-        return "PolygonInhom: " + str(list(zip(self.x, self.y)))
+        return "PolygonInhom: " + str(list(zip(self.x, self.y, strict=False)))
 
     def isinside(self, x, y):
         rv = 0
@@ -79,7 +79,7 @@ class PolygonInhom(AquiferData):
                 self.zcout[i].real, self.zcout[i].imag
             )
             if (aqout == self.model.aq) or (aqout.inhom_number > self.inhom_number):
-                ls = IntHeadDiffLineSink(
+                IntHeadDiffLineSink(
                     self.model,
                     x1=self.x[i],
                     y1=self.y[i],
@@ -93,7 +93,7 @@ class PolygonInhom(AquiferData):
                     aqin=aqin,
                     aqout=aqout,
                 )
-                ls = IntFluxDiffLineSink(
+                IntFluxDiffLineSink(
                     self.model,
                     x1=self.x[i],
                     y1=self.y[i],
@@ -171,8 +171,8 @@ class PolygonInhomMaq(PolygonInhom):
         model,
         xy,
         kaq=1,
-        z=[1, 0],
-        c=[],
+        z=None,
+        c=None,
         npor=0.3,
         topboundary="conf",
         hstar=None,
@@ -180,6 +180,10 @@ class PolygonInhomMaq(PolygonInhom):
         order=3,
         ndeg=3,
     ):
+        if c is None:
+            c = []
+        if z is None:
+            z = [1, 0]
         if N is not None:
             assert (
                 topboundary[:4] == "conf"
@@ -197,8 +201,9 @@ class PolygonInhomMaq(PolygonInhom):
 
 
 class PolygonInhom3D(PolygonInhom):
-    """Model3D Class to create a multi-layer model object consisting of many aquifer
-    layers. The resistance between the layers is computed from the vertical hydraulic
+    """Create a multi-layer model object consisting of many aquifer layers.
+
+    The resistance between the layers is computed from the vertical hydraulic
     conductivity of the layers.
 
     Parameters
@@ -251,7 +256,7 @@ class PolygonInhom3D(PolygonInhom):
         model,
         xy,
         kaq=1,
-        z=[1, 0],
+        z=None,
         kzoverkh=1,
         npor=0.3,
         topboundary="conf",
@@ -262,6 +267,8 @@ class PolygonInhom3D(PolygonInhom):
         order=3,
         ndeg=3,
     ):
+        if z is None:
+            z = [1, 0]
         if N is not None:
             assert (
                 topboundary[:4] == "conf"
@@ -277,7 +284,7 @@ class PolygonInhom3D(PolygonInhom):
 
 def compute_z1z2(xy):
     # Returns z1 and z2 of polygon, in clockwise order
-    x, y = list(zip(*xy))
+    x, y = list(zip(*xy, strict=False))
     if x[0] == x[-1] and y[0] == y[-1]:  # In case last point is repeated
         x = x[:-1]
         y = y[:-1]
@@ -311,7 +318,7 @@ class BuildingPit(AquiferData):
         npor=0.3,
         order=3,
         ndeg=3,
-        layers=[0],
+        layers=None,
     ):
         """Element to simulate a building pit with an impermeable wall.
 
@@ -355,6 +362,8 @@ class BuildingPit(AquiferData):
         layers: list or np.array
             layers in which impermeable wall is present.
         """
+        if layers is None:
+            layers = [0]
         AquiferData.__init__(self, model, kaq, c, z, npor, ltype)
         self.order = order
         self.ndeg = ndeg
@@ -501,14 +510,14 @@ class BuildingPitMaq(BuildingPit):
         model,
         xy,
         kaq=1.0,
-        c=[],
-        z=[1, 0],
+        c=None,
+        z=None,
         topboundary="conf",
         hstar=None,
         npor=0.3,
         order=3,
         ndeg=3,
-        layers=[0],
+        layers=None,
     ):
         """Element to simulate a building pit with an impermeable wall in ModelMaq.
 
@@ -554,6 +563,12 @@ class BuildingPitMaq(BuildingPit):
         layers: list or np.array
             layers in which impermeable wall is present.
         """
+        if layers is None:
+            layers = [0]
+        if z is None:
+            z = [1, 0]
+        if c is None:
+            c = []
         (kaq, c, npor, ltype) = param_maq(kaq, z, c, npor, topboundary)
         super().__init__(
             model=model,
@@ -577,7 +592,7 @@ class BuildingPit3D(BuildingPit):
         xy,
         kaq=1.0,
         kzoverkh=1.0,
-        z=[1, 0],
+        z=None,
         topboundary="conf",
         topres=0,
         topthick=0,
@@ -585,7 +600,7 @@ class BuildingPit3D(BuildingPit):
         npor=0.3,
         order=3,
         ndeg=3,
-        layers=[0],
+        layers=None,
     ):
         """Element to simulate a building pit with an impermeable wall in Model3D.
 
@@ -633,6 +648,10 @@ class BuildingPit3D(BuildingPit):
         layers: list or np.array
             layers in which impermeable wall is present.
         """
+        if layers is None:
+            layers = [0]
+        if z is None:
+            z = [1, 0]
         (kaq, c, npor, ltype) = param_3d(kaq, z, kzoverkh, npor, topboundary, topres)
         if topboundary == "semi":
             z = np.hstack((z[0] + topthick, z))
@@ -664,7 +683,7 @@ class LeakyBuildingPit(BuildingPit):
         hstar=None,
         order=3,
         ndeg=3,
-        layers=[0],
+        layers=None,
         res=np.inf,
     ):
         """Element to simulate a building pit with a leaky wall.
@@ -713,6 +732,8 @@ class LeakyBuildingPit(BuildingPit):
             shape (n_segments,) or (n_layers, n_segments). Default is np.inf,
             which simulates an impermeable wall.
         """
+        if layers is None:
+            layers = [0]
 
         super().__init__(
             model,
@@ -744,7 +765,9 @@ class LeakyBuildingPit(BuildingPit):
         if np.any(self.res < self.tiny):
             warn(
                 f"Found resistances smaller than {self.tiny}, "
-                f"these were replaced by {self.tiny}."
+                f"these were replaced by {self.tiny}.",
+                category=UserWarning,
+                stacklevel=1,
             )
             self.res[self.res < self.tiny] = self.tiny
 
@@ -895,16 +918,22 @@ class LeakyBuildingPitMaq(LeakyBuildingPit):
         model,
         xy,
         kaq=1.0,
-        z=[1, 0],
-        c=[],
+        z=None,
+        c=None,
         npor=0.3,
         topboundary="conf",
         hstar=None,
         order=3,
         ndeg=3,
-        layers=[0],
+        layers=None,
         res=np.inf,
     ):
+        if layers is None:
+            layers = [0]
+        if c is None:
+            c = []
+        if z is None:
+            z = [1, 0]
         (kaq, c, npor, ltype) = param_maq(kaq, z, c, npor, topboundary)
         super().__init__(
             model=model,
@@ -928,7 +957,7 @@ class LeakyBuildingPit3D(LeakyBuildingPit):
         model,
         xy,
         kaq=1.0,
-        z=[1, 0],
+        z=None,
         kzoverkh=1.0,
         npor=0.3,
         topboundary="conf",
@@ -937,7 +966,7 @@ class LeakyBuildingPit3D(LeakyBuildingPit):
         hstar=None,
         order=3,
         ndeg=3,
-        layers=[0],
+        layers=None,
         res=np.inf,
     ):
         """Element to simulate a building pit with a leaky wall in Model3D.
@@ -990,6 +1019,10 @@ class LeakyBuildingPit3D(LeakyBuildingPit):
             shape (n_segments,) or (n_segments, n_layers). Default is np.inf,
             which simulates an impermeable wall.
         """
+        if layers is None:
+            layers = [0]
+        if z is None:
+            z = [1, 0]
         (kaq, c, npor, ltype) = param_3d(kaq, z, kzoverkh, npor, topboundary, topres)
         if topboundary == "semi":
             z = np.hstack((z[0] + topthick, z))
@@ -1016,7 +1049,9 @@ class AreaSinkInhom(Element):
         )
         self.N = N
         self.xc = xc  # x-center of area-sink
-        # self.nparam = 1  # Defined here and not in Element as other elements can have multiple parameters per layers
+        # Defined here and not in Element as other elements can have
+        # multiple parameters per layers:
+        # self.nparam = 1
         # self.nunknowns = 0
         self.aq = aq
         self.model.add_element(self)
