@@ -178,7 +178,6 @@ class ImpLineDoublet(LineDoubletHoBase, DisvecEquation):
 
     Parameters
     ----------
-
     model : Model object
         Model to which the element is added
     x1 : scalar
@@ -201,7 +200,6 @@ class ImpLineDoublet(LineDoubletHoBase, DisvecEquation):
 
     See Also
     --------
-
     :class:`.ImpLineDoubletString`
     """
 
@@ -251,7 +249,6 @@ class LeakyLineDoublet(LineDoubletHoBase, LeakyWallEquation):
 
     Parameters
     ----------
-
     model : Model object
         Model to which the element is added
     x1 : scalar
@@ -276,7 +273,6 @@ class LeakyLineDoublet(LineDoubletHoBase, LeakyWallEquation):
 
     See Also
     --------
-
     :class:`.LeakyLineDoubletString`
     """
 
@@ -381,7 +377,8 @@ class LineDoubletStringBase(Element):
                 self.ldlist[0].xc[0], self.ldlist[0].yc[0]
             )
         self.parameters = np.zeros((self.nparam, 1))
-        ## As parameters are only stored for the element not the list, we need to combine the following
+        # As parameters are only stored for the element not the list,
+        # we need to combine the following
         self.xc = np.array([ld.xc for ld in self.ldlist]).flatten()
         self.yc = np.array([ld.yc for ld in self.ldlist]).flatten()
         self.xcin = np.array([ld.xcin for ld in self.ldlist]).flatten()
@@ -422,7 +419,6 @@ class ImpLineDoubletString(LineDoubletStringBase, DisvecEquation):
 
     Parameters
     ----------
-
     model : Model object
         Model to which the element is added
     xy : array or list
@@ -440,11 +436,12 @@ class ImpLineDoubletString(LineDoubletStringBase, DisvecEquation):
 
     See Also
     --------
-
     :class:`.ImpLineDoublet`
     """
 
-    def __init__(self, model, xy=[(-1, 0), (1, 0)], layers=0, order=0, label=None):
+    def __init__(self, model, xy=None, layers=0, order=0, label=None):
+        if xy is None:
+            xy = [(-1, 0), (1, 0)]
         self.storeinput(inspect.currentframe())
         LineDoubletStringBase.__init__(
             self,
@@ -473,7 +470,6 @@ class LeakyLineDoubletString(LineDoubletStringBase, LeakyWallEquation):
 
     Parameters
     ----------
-
     model : Model object
         Model to which the element is added
     xy : array or list
@@ -493,13 +489,12 @@ class LeakyLineDoubletString(LineDoubletStringBase, LeakyWallEquation):
 
     See Also
     --------
-
     :class:`.ImpLineDoublet`
     """
 
-    def __init__(
-        self, model, xy=[(-1, 0), (1, 0)], res=np.inf, layers=0, order=0, label=None
-    ):
+    def __init__(self, model, xy=None, res=np.inf, layers=0, order=0, label=None):
+        if xy is None:
+            xy = [(-1, 0), (1, 0)]
         self.storeinput(inspect.currentframe())
         LineDoubletStringBase.__init__(
             self,
