@@ -121,10 +121,8 @@ class LineSink1D(LineSink1DBase, MscreenWellEquation):
     In case the line-sink is screened in multiple layers, the discharge is distributed
     over the layers such that the head inside the line-sink is constant in all layers.
 
-
     Parameters
     ----------
-
     model : Model object
         Model to which the element is added
     xls : scalar
@@ -135,9 +133,11 @@ class LineSink1D(LineSink1DBase, MscreenWellEquation):
         resistance of line-sink
     wh : scalar or str
         distance over which water enters line-sink
-        if 'H': (default) distance is equal to the thickness of the aquifer layer (when flow comes mainly from one side)
-        if '2H': distance is twice the thickness of the aquifer layer (when flow comes from both sides)
-        if scalar: the width of the stream that partially penetrates the aquifer layer
+        * if 'H': (default) distance is equal to the thickness of the aquifer layer
+          (when flow comes mainly from one side)
+        * if '2H': distance is twice the thickness of the aquifer layer
+          (when flow comes from both sides)
+        * if scalar: the width of the stream that partially penetrates the aquifer layer
     layers : scalar, list or array
         layer(s) in which element is placed
         if scalar: element is placed in this layer
@@ -179,7 +179,6 @@ class HeadLineSink1D(LineSink1DBase, HeadEquation):
 
     Parameters
     ----------
-
     model : Model object
         Model to which the element is added
     xls : scalar
@@ -190,9 +189,11 @@ class HeadLineSink1D(LineSink1DBase, HeadEquation):
         resistance of line-sink
     wh : scalar or str
         distance over which water enters line-sink
-        if 'H': (default) distance is equal to the thickness of the aquifer layer (when flow comes mainly from one side)
-        if '2H': distance is twice the thickness of the aquifer layer (when flow comes from both sides)
-        if scalar: the width of the stream that partially penetrates the aquifer layer
+        * if 'H': (default) distance is equal to the thickness of the aquifer layer
+          (when flow comes mainly from one side)
+        * if '2H': distance is twice the thickness of the aquifer layer
+          (when flow comes from both sides)
+        * if scalar: the width of the stream that partially penetrates the aquifer layer
     layers : scalar, list or array
         layer(s) in which element is placed
         if scalar: element is placed in this layer
@@ -228,7 +229,7 @@ class HeadLineSink1D(LineSink1DBase, HeadEquation):
 
 
 class HeadDiffLineSink1D(LineSink1DBase, HeadDiffEquation):
-    """HeadDiffLineSink1D for left side (xcout)"""
+    """HeadDiffLineSink1D for left side (xcout)."""
 
     def __init__(self, model, xls, label=None, aq=None, aqin=None, aqout=None):
         LineSink1DBase.__init__(
@@ -263,7 +264,7 @@ class HeadDiffLineSink1D(LineSink1DBase, HeadDiffEquation):
 
 
 class FluxDiffLineSink1D(LineSink1DBase, DisvecDiffEquation):
-    """HeadDiffLineSink1D for left side (xcout)"""
+    """HeadDiffLineSink1D for left side (xcout)."""
 
     def __init__(self, model, xls, label=None, aq=None, aqin=None, aqout=None):
         LineSink1DBase.__init__(
