@@ -2,7 +2,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 from matplotlib.collections import LineCollection
 
-from .trace import timtraceline, timtracelines
+from .trace import timtraceline
 
 plt.rcParams["contour.negative_linestyle"] = "solid"
 
@@ -24,7 +24,6 @@ class PlotTim:
 
         Parameters
         ----------
-
         win : list or tuple
             [xmin, xmax, ymin, ymax]
         newfig : boolean (default True)
@@ -41,10 +40,8 @@ class PlotTim:
 
         Returns
         -------
-
         None
         """
-
         if newfig:
             plt.figure(figsize=figsize)
             ax1 = None
@@ -120,7 +117,6 @@ class PlotTim:
 
         Parameters
         ----------
-
         win : list or tuple
             [xmin, xmax, ymin, ymax]
         ngr : scalar, tuple or list
@@ -149,10 +145,8 @@ class PlotTim:
 
         Returns
         -------
-
         cs : list of contour sets for each contoured layer
         """
-
         x1, x2, y1, y2 = win
         if np.isscalar(ngr):
             nx = ny = ngr
@@ -167,9 +161,9 @@ class PlotTim:
         # color
         if color is None:
             c = plt.rcParams["axes.prop_cycle"].by_key()["color"]
-        elif type(color) is str:
+        elif isinstance(color, str):
             c = len(layers) * [color]
-        elif type(color) is list:
+        elif isinstance(color, list):
             c = color
         if len(c) < len(layers):
             n = np.ceil(self.aq.naq / len(c))
@@ -185,7 +179,7 @@ class PlotTim:
             if labels:
                 fmt = "%1." + str(decimals) + "f"
                 plt.clabel(cs, fmt=fmt)
-        if type(legend) is list:
+        if isinstance(legend, list):
             plt.legend(cshandlelist, legend)
         elif legend:
             legendlist = ["layer " + str(i) for i in layers]
@@ -213,7 +207,6 @@ class PlotTim:
 
         Parameters
         ----------
-
         win : list or tuple
             [xmin, xmax, ymin, ymax]
         n : integer
@@ -240,10 +233,8 @@ class PlotTim:
 
         Returns
         -------
-
         cs : contour set
         """
-
         x1, x2, y1, y2 = win
         h = self.headalongline(
             np.linspace(x1 + nudge, x2 - nudge, n),
@@ -283,7 +274,7 @@ class PlotTim:
         silent=".",
         color=None,
         orientation="hor",
-        win=[-1e30, 1e30, -1e30, 1e30],
+        win=None,
         newfig=False,
         figsize=None,
         *,
@@ -329,15 +320,16 @@ class PlotTim:
 
         Returns
         -------
-
         traces : result
             only if return_traces = True
         """
+        if win is None:
+            win = [-1e30, 1e30, -1e30, 1e30]
         if color is None:
             c = plt.rcParams["axes.prop_cycle"].by_key()["color"]
-        elif type(color) is str:
+        elif isinstance(color, str):
             c = self.aq.naq * [color]
-        elif type(color) is list:
+        elif isinstance(color, list):
             c = color
         if len(c) < self.aq.naq:
             n = int(np.ceil(self.aq.naq / len(c)))
@@ -427,7 +419,6 @@ class PlotTim:
 
         Parameters
         ----------
-
         x1 : scalar
             left edge of contour domain
         x2 : scalar
@@ -455,7 +446,6 @@ class PlotTim:
 
         Returns
         -------
-
         ax : axis
         """
         naq = self.aq.naq
